+++ R/functions.R
@@ -46,7 +46,7 @@ fishtree_phylogeny <- function(species, rank, type = c("chronogram", "phylogram"
     valid_names <- .name_check(species)
     if (length(valid_names) < 2) rlang::abort("Must include at least 2 sampled tips in `species`")
     tree <- .get(fullurl, ape::read.tree)
-    return(ape::drop.tip(tree, tree$tip.label[!tree$tip.label %in% gsub(" ", "_", valid_names)]))
+    return(ape::keep.tip(tree, tree$tip.label[tree$tip.label %in% gsub(" ", "_", valid_names)]))
   }
 
   res <- .fetch_rank(rank)
@@ -123,8 +123,9 @@ fishtree_rogues <- function(rank) {
 #' paste("There are", n_sampl, "sampled species out of", n_total, "in wrasses.")
 fishtree_taxonomy <- function(ranks = NULL) {
   tax <- .get("https://fishtreeoflife.org/api/taxonomy.json", jsonlite::fromJSON)
-  tax_df <- .list2df(tax)
-  colnames(tax_df) <- c("rank", "name")
+  tax_df <- utils::stack(tax)
+  colnames(tax_df) <- c("name", "rank")
+  tax_df <- tax_df[c("rank", "name")]
   if (is.null(ranks)) return(tax_df)
 
   wanted <- tax_df[tax_df$name %in% ranks, ]
@@ -166,7 +167,6 @@ fishtree_alignment <- function(species, rank, split = FALSE) {
 
   if (rlang::is_missing(rank)) {
     url <- "https://fishtreeoflife.org/downloads/final_alignment.phylip.xz"
-    if(!exists(url, envir = .cache)) rlang::inform("Loading alignment for all species, this will take a while...")
     nlines <- 11650
   } else {
     res <- .fetch_rank(rank)
@@ -252,4 +252,54 @@ fishtree_tip_rates <- function(species, rank, sampled_only = TRUE) {
     if (rlang::is_empty(wanted)) rlang::abort(paste("Can't get tip rates for", what, rank, "because there are not enough sampled species"))
     return(rates[rates$species %in% wanted, ])
   }
+}
+
+
+
+
+
+
+
+#' Get complete (stochastically-resolved) phylogenies from the Fish Tree of Life
+#'
+#' Retrieves a complete, stochastically-resolved phylogeny via the Fish Tree of Life API. If neither `species` nor `rank` are specified, returns the entire phylogeny.
+#'
+#' @inheritParams fishtree_phylogeny
+#' @param mc.cores Number of cores to use in \link[parallel]{mclapply} when subsetting the tree (default `1`)
+#' @return An object of class `"multiPhylo"`.
+#' @export
+#' @references
+#' Rabosky, D. L., Chang, J., Title, P. O., Cowman, P. F., Sallan, L., Friedman, M., Kashner, K., Garilao, C., Near, T. J., Coll, M., Alfaro, M. E. (2018). An inverse latitudinal gradient in speciation rate for marine fishes. Nature, 559(7714), 392–395. doi:10.1038/s41586-018-0273-1
+#'
+#' Enhanced polytomy resolution strengthens evidence for global gradient in speciation rate for marine fishes. \url{https://fishtreeoflife.org/rabosky-et-al-2018-update/}
+#' @examples
+#' \dontrun{
+#' tree <- fishtree_complete_phylogeny(rank = "Acanthuridae")
+#' sampled_tips <- fishtree_phylogeny(rank = "Acanthuridae")$tip.label
+#' all_tips <- tree[[1]]$tip.label
+#' new_tips <- setdiff(all_tips, sampled_tips)
+#' par(mfrow = c(2,2))
+#' for (ii in 1:4) {
+#'   plot(tree[[ii]], show.tip.label = FALSE, no.margin = TRUE)
+#'   ape::tiplabels(pch = 19, col = ifelse(tree[[ii]]$tip.label %in% new_tips, "red", NA))
+#' }
+#' }
+#' @import parallel
+fishtree_complete_phylogeny <- function(species, rank, mc.cores = getOption("mc.cores", 1L)) {
+  if (!rlang::is_missing(species) && !rlang::is_missing(rank)) rlang::abort("Must supply at most one of either `species` or `rank`, not both")
+
+  trees <- .get("https://fishtreeoflife.org/downloads/actinopt_full.trees.xz", ape::read.tree)
+  if (rlang::is_missing(species) && rlang::is_missing(rank)) return(trees)
+
+  if (!rlang::is_missing(rank)) {
+    res <- .fetch_rank(rank)
+    if (length(res[[1]][[1]]$rogues) > 0) rlang::warn(paste(res[[2]], rank, "is not monophyletic; only including species in this taxon"))
+    valid_spp <- .name_check(res[[1]][[1]]$species, trees[[1]]$tip.label)
+  } else if (!rlang::is_missing(species)) {
+    valid_spp <- .name_check(species, trees[[1]]$tip.label)
+  }
+  valid_spp <- gsub(" ", "_", valid_spp) # fix up tip names
+  res <- parallel::mclapply(trees, ape::keep.tip, tip = valid_spp, mc.cores = mc.cores)
+  class(res) <- "multiPhylo"
+  res
 }