+++ NEWS.md
@@ -1,3 +1,17 @@
+# fishtree 0.3.0
+
+# fishtree (development version)
+
+* New function: `fishtree_complete_phylogeny`, gets complete phylogenies
+  where unsampled taxa were placed using stochastic polytomy resolution.
+
+* Use package `memoise` instead of rolling our own caching functionality.
+
+* New global option `fishtree.quiet`: when set to `FALSE`, be more verbose
+  during downloads (can help with debugging issues).
+
+* Some internal reorganization
+
 # fishtree 0.2.0
 
 * New function: `fishtree_rogues`, identifies rogue/intruder taxa that break